Indicate that login failed due to unverified email address
Previously whenever login failed (whether due to incorrect email/password or due to unverified email address)
the response always said {login:false}. This makes it difficult for a front-end/web/mobile application to understand
whether to tell the user to retry or logging in or whether to tell the user to ensure he has verified his email address.

With these changes, hopefully it will be clear for the front-end/mobile/web app to understand what is the problem.